Refactor: 결제페이지 구조 변경
