+++ package.json
@@ -19,6 +19,7 @@
     "@mantine/core": "^7.12.1",
     "@mantine/dates": "^7.12.1",
     "@mantine/hooks": "^7.12.1",
+    "@paypal/react-paypal-js": "^8.6.0",
     "@types/react-router-dom": "^5.3.3",
     "autoprefixer": "^10.4.20",
     "axios": "^1.3.1",
